Move Game.css and Game.js to Tracking folder
